[NB] Add `place` and `route` methods to `cMain`...
Add `place` and `route` methods to `cMain` Cython extension class.

The `place` method returns an array of block-positions, with dimensions
indexed as follows:

    block-index, x=0/y=1/slot-index=2

The `route` method returns a `collections.OrderedDict` containing
details of the routing results.  __NB__ The results returned do not
contain the actual routing result, but only contains information about
timing of the resultant routings, including critical-path-delay.  The
benefit of this is that we can ignore the routing output file when we
are only concerned with the delay information.  Since the routing output
is quite large _(on the order of MB)_, this can save _a lot_ of space.