[NB] Add generator API to BoundFinderBase Pytho...
Add generator API to `BoundFinderBase` Python class.  This allows for
custom handling to be performed on each search iteration, if necessary.

To use the generator interface, set the `generator` keyword argument to
`True` when calling the `BoundFinderBase.start` method.  The `start`
method will then return a `generator` instance, which, will in turn
return a generator instance until the boundary-of-interest is reached.
Therefore, the return type can be checked to determine if the boundary
has been reached, where a return value of type `types.GeneratorType`
means the search has not finished yet.  Otherwise, the return value is
the boundary-of-interest.

For example usage, see the `BoundFinderBase._run` method.`