Add cyvpr.bin.do_route.route function, save ro...
Add `cyvpr.bin.do_route.route` function, which saves route-results in a HDF
file, with one table row per-route-state.  Each table row contains information
about the parameters used for the route, including channel-width, as well as if
the route was successful.  In addition, the following attributes are available
for each net: the number of bends, the wire-length, and the number of
wire-segments.

__NB__ The new `route` function is not yet the default action when running
`cyvpr.bin.do_route.route` as a main script.