[WIP] Add code/docs to find remaining routings...
Add code and documentation to the `Routing results` IPython notebook
describing how to find the remaining routings to be performed for all
channel-widths where:

  * At least one placement was routed successfully and at least one was
    not.
  * The channel-width is the minimum where all placements are
    successfully routed.

By performing all such routings, a great amount of flexibility is given
in regards to how the results are interpreted.  Since channel-width is a
fixed parameter for a target architecture, the minimum channel-width may
not always be the primary objective.  However, by recording information
about what channel-width is required for each placement, we can get an
idea of the "routability" of a placer relative to other placers.

The remaining routing tasks is given in the form of a list of
`(block_positions_sha1, width_fac)` pairs.