Add notebook plotting significance of critical-...
Add IPython notebook plotting the routing critical-path-delay results from
running 10 trials for each of the following place algorithms:

 * `vpr-not_fast-bounding_box`
 * `vpr-fast-bounding_box`
 * `cyplace-7a8d71d4`

Note the _$p$-value_ significance markings on the plots.  The significance
_$p$-values_ are computed using a [Wilcoxon signed-rank test][1] and are
labelled on the plot above either a _red_ or _green_ arrow, pointing to the
relevant means in the box plot.

A _green_ arrow indicates that the $p$-value is less than 0.05, such that the
null-hypothesis may be rejected, indicating that the means are not
statistically likely to be due to drawing from the same population.  If the
$p$-value is greater than 0.05, the arrow is _red_.

[1]: http://en.wikipedia.org/wiki/Wilcoxon_signed-rank_test