Add HDF label customization to plot_place_stats...
The label of any HDF results file _may_ be customized by prepending a label
followed by `::` to the corresponding `hdf5_placements_file` path.

For example, when plotting:

    python -i -m cyvpr.bin.plot_place_stats 'tseng' cyplace::placed-results.h5 --plot-linewidth=1

The results-file `placed-results.h5` will be assigned the label `'cyplace'`.
If the prefix was not provided, the results-file namebase would be used instead
_(i.e., `placed-results` in this case)_.