Use GridSpec in bin.plot_place_stats.plot_h5fs...
Use `matplotlib.gridspec.GridSpec` in `bin.plot_place_stats.plot_h5fs` to
specify layout of trend-graphs and place-stats box-plots.  This decouples the
contents of the trend-graphs area and the box-plots area, such that the number
of rows and columns in each area _(i.e., trend-graphs or box-plots)_ can be
changed independently of the configuration of the opposite area.

Recall that the available layout configurations are:

`None`
======
    _____________
    |           |
    |   trends  |
    |___________|

`top`
=====
    _____________
    |  |  |  |  |
    | S| T| A| T|
    |__|__|__|__|
    |           |
    |   trends  |
    |___________|

`bottom`
========
    _____________
    |           |
    |   trends  |
    |___________|
    |  |  |  |  |
    | S| T| A| T|
    |__|__|__|__|

`left`
======
    _________________________
    |  |  |  |  |           |
    | S| T| A| T|   trends  |
    |__|__|__|__|___________|

`right`
=======
    _________________________
    |           |  |  |  |  |
    |   trends  | S| T| A| T|
    |___________|__|__|__|__|